requirements.txt
numpy>=1.26,<3.0
pandas>=2.2,<3.0
pyarrow>=16.0,<22.0
requests>=2.32,<3.0
scikit-learn>=1.5,<2.0
matplotlib>=3.9,<4.0
python-dotenv>=1.0,<2.0
torch>=2.3,<3.0
jupyter>=1.0,<2.0
python-pptx>=0.6.23,<1.0
streamlit>=1.44,<2.0
